style: pack middleware to exclusive package
