+++ gateway/mw/cors.go
@@ -1,0 +1,35 @@
+/*
+Copyright 2024 The west2-online Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package mw
+
+import (
+	"github.com/cloudwego/hertz/pkg/app"
+	"github.com/hertz-contrib/cors"
+
+	"github.com/west2-online/DomTok/pkg/constants"
+)
+
+func CorsMW() app.HandlerFunc {
+	return cors.New(cors.Config{
+		AllowOrigins:     []string{"*"},
+		AllowCredentials: true,
+		AllowMethods:     []string{"GET", "POST", "PUT", "DELETE", "OPTIONS"},
+		AllowHeaders:     []string{"*"},
+		MaxAge:           constants.CorsMaxAge,
+		ExposeHeaders:    []string{"Content-Length"},
+	})
+}
